Split error classes list into two files
